Relace TimeTable component with Table
