--- frontend/src/components/TimeTable/TimeTable.scss
@@ -1,57 +1,0 @@
-
-// Define the styles for the timetable container
-.timetable {
-  height: 100%;
-  width: 100%;
-  box-shadow: 0 0 10px rgba(0, 0, 0, 0.2);
-  // Define the styles for the timetable header
-    thead {
-        width: 100%;
-        font-size: 10px;
-        border-radius: 10px 10px 0 0;
-        box-shadow: 0 2px 5px rgba(0, 0, 0, 0.1);
-        th {
-            padding: 1em;
-            border-bottom: 1px solid black;
-            // Add left border for second th child and first child
-            &:nth-child(2), &:first-child {
-                border-left: 1px solid black;
-            }
-            //right border for last th child
-            &:last-child {
-                border-right: 1px solid black;
-            }
-        }
-    }
-    tbody {
-        font-size: 12px;
-        font-weight: 500;
-        // Fit the first td to content
-        td {
-            padding: 1em 0;
-            // Set even width for all td
-            width: 14.28%;
-            text-align: center;
-            margin: 0;
-            // Add bottom border
-            border-bottom: 1px solid #D9D9D9;
-            &.course{
-                border: 1px solid #D9D9D9;
-                background-color: #E8F1FA;
-            }
-            &:first-child {
-                width: 5%;
-                text-align: center;
-                // Only set left, right and bottom border
-                border-left: 1px solid black;
-                border-right: 1px solid black;
-                border-bottom: 1px solid black;
-            }
-            &.course {
-                background-color: var(--primary-color);
-            }
-        }
-    }
-}
-
-
