+++ frontend/src/pages/TimeTablePage/TimeTablePage.tsx
@@ -2,18 +2,70 @@ import './TimeTablePage.scss'
 import React, { useState } from 'react';
 import filter_icon from 'assets/images/filter-icon.svg';
 import caret_down from 'assets/images/caret-down-solid.svg';
-import TimeTable from 'components/TimeTable/TimeTable';
+import Table, { Row, TableType } from 'components/Table/Table';
 
-interface Course {
+type Course = {
     id: string;
     name: string;
     credits: number;
     slots: number;
     description: string;
     lecturer: string;
+    day: string;
     start: number;
-    end: number;
-}
+    duration: number;
+  }
+
+const days = ['Mon', 'Tue', 'Wed', 'Thu', 'Fri', 'Sat', 'Sun'];
+
+const rowGenerator = (courses: Course[]) => {
+    // Create emtpy rows which has 8 columns, the first one is the order of the row
+    // There are 10 rows in total
+    let rows: Row[]= [];
+    for (let i = 1; i <= 10; i++)
+    {
+      // Generate an empty row
+      const row = new Row();
+      // Iterate through each day (col)
+      // If there is a course that matches the day and start time, add it to the row
+      days.forEach(day => {
+        const found = courses.find(course => course.day === day && course.start === i)
+        if (found) {
+          row.cols.push({ name: found.name, rowSpan: found.duration })
+        } else {
+          row.cols.push({ name: '', rowSpan: 1 })
+        }
+      })
+  
+      rows.push(row);
+    }
+    return rows
+  }
+
+const courses: Course[] = [
+{
+    id: '1',
+    name: 'Probability',
+    credits: 3,
+    slots: 2,
+    description: 'Probability and Mathematical Statistics',
+    lecturer: 'Nguyen Van B',
+    day: 'Tue',
+    start: 1,
+    duration: 3,
+},
+{
+    id: '2',
+    name: 'Probability',
+    credits: 3,
+    slots: 2,
+    description: 'Probability and Mathematical Statistics',
+    lecturer: 'Nguyen Van B',
+    day: 'Tue',
+    start: 5,
+    duration: 3,
+},
+]
 
 function TimeTablePage() {
     // create an array of days in a week with abbreviation
@@ -27,7 +79,7 @@ function TimeTablePage() {
         event.preventDefault();
         // TODO: Fetch courses based on filters
     };
-    
+
     return (
         <div className="timetable-page page">
             <form onSubmit={handleFilter} className='input-form'>
@@ -45,7 +97,10 @@ function TimeTablePage() {
                     Filter<img src={filter_icon} alt="test" />
                 </button>
             </form>
-            <TimeTable />
+            <Table 
+                headers={days} 
+                type={TableType.TimeTable} 
+                contents={rowGenerator(courses)}/>
         </div>
     );
 }
